ref(tiller): refactor tiller APIs into logical files
*ReleaseServer.GetReleaseContent -> tiller/release_content.go
*ReleaseServer.GetReleaseStatus -> tiller/release_status.go
