+++ pkg/tiller/release_server.go
@@ -220,70 +220,6 @@ func filterReleases(filter string, rels []*release.Release) ([]*release.Release,
 	return matches, nil
 }
 
-// GetReleaseStatus gets the status information for a named release.
-func (s *ReleaseServer) GetReleaseStatus(c ctx.Context, req *services.GetReleaseStatusRequest) (*services.GetReleaseStatusResponse, error) {
-	if !ValidName.MatchString(req.Name) {
-		return nil, errMissingRelease
-	}
-
-	var rel *release.Release
-
-	if req.Version <= 0 {
-		var err error
-		rel, err = s.env.Releases.Last(req.Name)
-		if err != nil {
-			return nil, fmt.Errorf("getting deployed release %q: %s", req.Name, err)
-		}
-	} else {
-		var err error
-		if rel, err = s.env.Releases.Get(req.Name, req.Version); err != nil {
-			return nil, fmt.Errorf("getting release '%s' (v%d): %s", req.Name, req.Version, err)
-		}
-	}
-
-	if rel.Info == nil {
-		return nil, errors.New("release info is missing")
-	}
-	if rel.Chart == nil {
-		return nil, errors.New("release chart is missing")
-	}
-
-	sc := rel.Info.Status.Code
-	statusResp := &services.GetReleaseStatusResponse{
-		Name:      rel.Name,
-		Namespace: rel.Namespace,
-		Info:      rel.Info,
-	}
-
-	// Ok, we got the status of the release as we had jotted down, now we need to match the
-	// manifest we stashed away with reality from the cluster.
-	resp, err := s.ReleaseModule.Status(rel, req, s.env)
-	if sc == release.Status_DELETED || sc == release.Status_FAILED {
-		// Skip errors if this is already deleted or failed.
-		return statusResp, nil
-	} else if err != nil {
-		s.Log("warning: Get for %s failed: %v", rel.Name, err)
-		return nil, err
-	}
-	rel.Info.Status.Resources = resp
-	return statusResp, nil
-}
-
-// GetReleaseContent gets all of the stored information for the given release.
-func (s *ReleaseServer) GetReleaseContent(c ctx.Context, req *services.GetReleaseContentRequest) (*services.GetReleaseContentResponse, error) {
-	if !ValidName.MatchString(req.Name) {
-		return nil, errMissingRelease
-	}
-
-	if req.Version <= 0 {
-		rel, err := s.env.Releases.Deployed(req.Name)
-		return &services.GetReleaseContentResponse{Release: rel}, err
-	}
-
-	rel, err := s.env.Releases.Get(req.Name, req.Version)
-	return &services.GetReleaseContentResponse{Release: rel}, err
-}
-
 // UpdateRelease takes an existing release and new information, and upgrades the release.
 func (s *ReleaseServer) UpdateRelease(c ctx.Context, req *services.UpdateReleaseRequest) (*services.UpdateReleaseResponse, error) {
 	err := s.env.Releases.LockRelease(req.Name)
